+++ src/variables/SeomaticVariable.php
@@ -146,7 +146,7 @@ class SeomaticVariable extends ServiceLocator
      */
     public function getEnvironment()
     {
-        return Seomatic::$environment;
+        return Craft::parseEnv(Seomatic::$environment);
     }
 
     /**
@@ -163,6 +163,10 @@ class SeomaticVariable extends ServiceLocator
         }
         $env = Seomatic::$environment;
         $settingsUrl = UrlHelper::cpUrl('seomatic/plugin');
+        $general = Craft::$app->getConfig()->getGeneral();
+        if (!$general->allowAdminChanges) {
+            $settingsUrl = '';
+        }
         // If they've manually overridden the environment, just return it
         if (EnvironmentHelper::environmentOverriddenByConfig()) {
             return Craft::t('seomatic', 'This is overridden by the `config/seomatic.php` config setting',
@@ -177,6 +181,9 @@ class SeomaticVariable extends ServiceLocator
         }
         // Try to also check the `ENVIRONMENT` env var
         $envVar = getenv('ENVIRONMENT');
+        if (Seomatic::$settings->manuallySetEnvironment) {
+            $envVar = $settingsEnv;
+        }
         if (!empty($envVar)) {
             $env = EnvironmentHelper::determineEnvironment();
             switch ($env) {