+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # SEOmatic Changelog
 
+## 3.3.33 - 2021.02.23
+### Added
+* Added a **Manually Set SEOmatic Environment** setting (defaults to off) for people who want to manually set the SEOmatic environment via environment variable
+
+### Changed
+* Updated build system infra
+* Parse the SEOmatic Environment as a variable before displaying it on the dashboard
+
+### Fixed
+* No longer include any `<body>` script tags (as in Google Tag Manager) if the entire script meta container is excluded
+* The `og:locale` tag no longer shows the wrong language for headless requests
+* Don't link to the settings page if `allowAdminChanges` is `false`
+* No longer URL decode text that is pulled from another field in SEOmatic
+
 ## 3.3.32 - 2021.02.12
 ### Fixed
 * Fixed a typo in the global JsonLDContainer that caused errors to be generated